+++ CHANGELOG.md
@@ -18,6 +18,7 @@
     - Add `evalString`. [#30](https://github.com/DiscreteTom/retsac/issues/30)
     - Add `comment`.
     - **_Breaking Change_**: move `regexLiteral/numericLiteral` into `Lexer.javascript`.
+  - Feat: add option `lineContinuation` for `Lexer.stringLiteral`.
   - Perf: accelerate expectational lexing by using `Map` to select expected definitions.
 - Optimize package size.